requirements.txt
aiohttp
opencv-python
PyQt5
Pillow